Implement CPU unpacking for GPU YUV422 output

+++ include/Graphics/GpuYuvConverter.h
@@ -17,7 +17,7 @@ public:
     void cleanup();
 
     bool convertAndReadback(const uint16_t* raw, int width, int height,
-                            std::vector<uint16_t>& outPacked);
+                            std::vector<uint32_t>& outPacked);
 private:
     Renderer_VK* m_renderer;
